Add new version of Dijkstra's SP

+++ src/dataStructures/digraphs/DirectedEdge.java
@@ -26,7 +26,7 @@ public class DirectedEdge {
         return w;
     }
 
-    public double weight() {
+    public int weight() {
         return weight;
     }
